Correções service e repository

+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/afastamento/AfastamentoRepository.java
@@ -4,6 +4,6 @@ import com.sistemaRegistroVerificacao.model.entity.afastamento.Afastamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AfastamentoRepository extends JpaRepository<Afastamento, Long>, JpaSpecificationExecutor<Afastamento> {
+public interface AfastamentoRepository extends JpaRepository<Afastamento, Integer>, JpaSpecificationExecutor<Afastamento> {
 }
