+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/atividade/AtividadeRepository.java
@@ -4,7 +4,7 @@ import com.sistemaRegistroVerificacao.model.entity.atividade.Atividade;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface AtividadeRepository extends JpaRepository<Atividade, Long>, JpaSpecificationExecutor<Atividade> {
+public interface AtividadeRepository extends JpaRepository<Atividade, Integer>, JpaSpecificationExecutor<Atividade> {
 }
