+++ src/main/java/com/sistemaRegistroVerificacao/model/repository/sala/SalaRepository.java
@@ -4,6 +4,6 @@ import com.sistemaRegistroVerificacao.model.entity.sala.Sala;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface SalaRepository extends JpaRepository<Sala, Long>, JpaSpecificationExecutor<Sala> {
+public interface SalaRepository extends JpaRepository<Sala, Integer>, JpaSpecificationExecutor<Sala> {
 }
